Renamed _dagcircuit to dagcircuit
