+++ qiskit/dagcircuit/__init__.py
@@ -6,6 +6,6 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """Module for DAG Circuits."""
-from ._dagcircuit import DAGCircuit
-from ._dagnode import DAGNode
+from .dagcircuit import DAGCircuit
+from .dagnode import DAGNode 
 from .exceptions import DAGCircuitError
