+++ qiskit/dagcircuit/dagcircuit.py
